Adding dataset-related classes.
